requirements.txt
# Shadow — Python Dependencies
# Install: pip install -r requirements.txt

# --- Core ---
openai              # Ollama-compatible API client
anthropic           # Apex escalation to Claude API
pyyaml              # Config file loading
pydantic-settings>=2.5,<3   # Centralized typed config (shadow.config) — pulls pydantic v2
python-dotenv       # Environment variables (used internally by pydantic-settings)
tzdata              # Timezone data (required on Windows, harmless on Linux)

# --- Memory (Grimoire) ---
chromadb            # Vector database for semantic search

# --- Research (Reaper) ---
ddgs                # DDG search (formerly duckduckgo-search)
beautifulsoup4      # Web scraping
requests            # HTTP client
aiohttp             # Async HTTP client
playwright          # Stealth browser automation
yt-dlp              # YouTube subtitle download
praw                # Reddit API

# --- Scheduling ---
apscheduler         # Background task scheduler (standing tasks)

# --- Monitoring ---
psutil              # System metrics (CPU, RAM, disk, process)
watchdog            # File system monitoring
rich                # Terminal formatting

# --- API (future) ---
fastapi             # REST API server
uvicorn             # ASGI server

# --- Testing ---
pytest              # Test framework
pytest-asyncio      # Async test support
httpx               # Required by fastapi TestClient
